requirements.txt
natsort==8.1.0
pytest==7.0.1
requests==2.23.0
pandas==1.2.4
PyYAML==5.4.1